+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The following sections are considered for each release: **Added, Changed, Fixed,
 ## [Unreleased]
 - Fix the order of comments on tickets so they are sorted sequentially.
 - Rebase Icon was showing 'STALE' text when task status was 'DIVERGED'
+### Fixed
+- Fix issue where term edits fail when the only change is case (uppercase to lowercase or vice versa).
+
 
 ## [1.3.35] - 2026-02-13