+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 The following sections are considered for each release: **Added, Changed, Fixed, Security, Deprecated, Removed**
 
-## [Unreleased]
+## [1.3.36] - 2026-02-23
 - Fix the order of comments on tickets so they are sorted sequentially.
 - Rebase Icon was showing 'STALE' text when task status was 'DIVERGED'
 - feature/show-ticket-authoring-number-on-ctpp-concept